+++ sdk/metric/internal/aggregate/filtered_reservoir_test.go
@@ -13,7 +13,6 @@ import (
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/sdk/metric/exemplar"
-	"go.opentelemetry.io/otel/sdk/metric/internal/reservoir"
 )
 
 func TestConcurrentSafeFilteredReservoir(t *testing.T) {
@@ -80,7 +79,7 @@ func (r *notConcurrentSafeReservoir) Collect(dest *[]exemplar.Exemplar) {
 type concurrentSafeReservoir struct {
 	base notConcurrentSafeReservoir
 	sync.Mutex
-	reservoir.ConcurrentSafe
+	exemplar.ConcurrentSafe
 }
 
 func (r *concurrentSafeReservoir) Offer(